Refactored structure of the src/ folder. Packages organised by the functionality.
