+++ src/pvtools/config/sensor_calibration_metrics.py
@@ -1,4 +1,5 @@
 import numpy as np
+
 from sklearn.metrics import mean_absolute_error, mean_squared_error, r2_score
 
 class SensorCalibrationMetrics:
